+++ app/package.json
@@ -94,6 +94,9 @@
             "^@pages(.*)$": "<rootDir>/src/pages$1",
             "^@services(.*)$": "<rootDir>/src/services$1",
             "^@utilities(.*)$": "<rootDir>/src/utilities$1"
-        }
+        },
+        "testMatch": [
+            "**/?(*.)+(test).[jt]s?(x)"
+        ]
     }
 }
